requirements.txt
streamlit
transformers
torch
pdfplumber
sentencepiece
streamlit-chat
Pillow               
